Multiple template proposal
Four templates demonstrating use of extension templates. The kitchensink will most likely be used to demonstrate new features as we add them. Right now it's a copy of my original API testbed. I will most like clean it up.

This is just a proposal. I'm okay with multiple repositories but I feel this approach is easier to manage.